+++ python/src/blaze_utils.cc
@@ -1,4 +1,5 @@
-// Copyright 2021 PANDA GmbH
+// Copyright 2021-2022 PANDA GmbH
+
 #include "blaze_utils.h"
 
 #include <utility>